# openapi.yaml
openapi: 3.0.0
info:
    title: 'Petrogassa API'
    description: >
        API de autenticación con NestJS (modelo cerrado, sin registro público):
        un admin da de alta usuarios por invitación; cada invitado activa su
        cuenta definiendo nombre, apellido y contraseña. Login con JWT en cookies
        HttpOnly, refresh tokens rotados (hasheados en DB, con detección de
        reutilización) y recuperación de contraseña. Todas las respuestas usan un
        sobre estándar (ver esquemas `SuccessEnvelope` / `ErrorResponse`).
    version: '0.0.1'
servers:
    - url: /api
      description: Servidor local
...
paths:
...
    /users:
        get:
            tags: [Users]
            summary: Listar usuarios (admin, paginado)
            description: >
                Grilla de administración de cuentas. Cada usuario incluye
                `status` (`pendiente` | `activo` | `desactivado`), derivado de
                `isActive` y de si la persona llegó a definir su contraseña.
                La contraseña nunca se expone.
            security: [{ cookieAuth: [] }]
            parameters:
                - in: query
                  name: search
                  description: Busca en nombre, apellido y email, ignorando mayúsculas y acentos.
                  schema: { type: string, maxLength: 100 }
                - { in: query, name: role, schema: { type: string, enum: [admin, rrhh, user] } }
                - { in: query, name: status, schema: { type: string, enum: [pendiente, activo, desactivado] } }
                - { in: query, name: page, schema: { type: integer } }
                - { in: query, name: limit, schema: { type: integer } }
            responses:
                '200': { description: Usuarios obtenidos. }
                '400': { $ref: '#/components/responses/ValidationError' }
                '401': { $ref: '#/components/responses/Unauthorized' }
    /users/{id}/revoke-sessions:
        post:
            tags: [Users]
            summary: Cerrar todas las sesiones de una cuenta (admin)
            description: >
                Revoca los refresh tokens del usuario e invalida sus access
                tokens ya emitidos, sin darlo de baja: la cuenta sigue activa
                (`status` no cambia) y puede volver a iniciar sesión con su
                contraseña. Pensado para una notebook perdida o una sesión
                abierta en una máquina ajena; para cortar el acceso por completo
                se usa `deactivate`. Un admin no puede cerrar las propias.
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '201':
                    description: Sesiones cerradas.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: object
                                        properties:
                                            sessionsClosed:
                                                type: integer
                                                description: Cuántas sesiones se cerraron.
                                                example: 2
                '401': { $ref: '#/components/responses/Unauthorized' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409':
                    description: Se intentó cerrar las sesiones propias.
    /users/{id}/deactivate:
        post:
            tags: [Users]
            summary: Quitar el acceso a una cuenta (admin)
            description: >
                Efecto inmediato: las sesiones abiertas dejan de funcionar en el
                siguiente request y se revocan sus refresh tokens. Un admin no
                puede desactivarse a sí mismo.
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '201': { description: Acceso desactivado. }
                '401': { $ref: '#/components/responses/Unauthorized' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409':
                    description: >
                        La cuenta ya no tiene acceso, o se intentó desactivar la
                        propia sesión.
    /users/{id}/reactivate:
        post:
            tags: [Users]
            summary: Devolver el acceso a una cuenta dada de baja (admin)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '201': { description: Acceso restablecido. }
                '401': { $ref: '#/components/responses/Unauthorized' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409':
                    description: >
                        La cuenta ya está activa, o nunca fue activada por su
                        titular (en ese caso corresponde reenviar la invitación).
    /users/{id}/roles:
        patch:
            tags: [Users]
            summary: Reemplazar los roles de una cuenta (admin)
            description: >
                Reemplaza el conjunto completo de roles; no es una modificación
                parcial. Toma efecto de inmediato, porque los roles se leen de
                la base en cada request. Un admin no puede cambiarse los suyos.
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [roles]
                            properties:
                                roles:
                                    type: array
                                    minItems: 1
                                    items: { type: string, enum: [admin, rrhh, user] }
                            example: { roles: [rrhh] }
            responses:
                '200': { description: Roles actualizados. }
                '400': { $ref: '#/components/responses/ValidationError' }
                '401': { $ref: '#/components/responses/Unauthorized' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409':
                    description: Se intentó cambiar los roles propios.
    /users/{id}:
        get:
            tags: [Users]
            summary: Obtener un usuario por id
            description: >
                Requiere sesión (cookie `accessToken`). Un usuario solo puede
                consultarse a sí mismo; un admin puede consultar a cualquiera.
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                '200':
                    description: Usuario obtenido.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessEnvelope'
                                    - type: object
                                      properties:
                                          message:
                                              example: 'Usuario obtenido correctamente'
                                          data:
                                              $ref: '#/components/schemas/UserEntity'
                '400':
                    $ref: '#/components/responses/ValidationError'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '429':
                    $ref: '#/components/responses/TooManyRequests'
        delete:
            tags: [Users]
            summary: Eliminar una invitación pendiente (admin)
            description: >
                Único borrado real de usuarios, acotado a cuentas con
                `status = pendiente`: una invitación que nunca se aceptó no
                tiene historia que preservar, y es la salida cuando se invitó a
                una dirección equivocada. Su token de activación se elimina en
                cascada y el email queda libre para reinvitarlo. Las cuentas en
                uso responden 409: esas se dan de baja con `deactivate`.
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '200': { description: Invitación eliminada. }
                '401': { $ref: '#/components/responses/Unauthorized' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409':
                    description: >
                        La cuenta ya fue activada alguna vez (activa o
                        desactivada); no se elimina, se desactiva.
...
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: accessToken
    responses:
        ValidationError:
            description: Error de validación (Bad Request)
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        Unauthorized:
            description: No autenticado / credenciales o token inválidos
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        Forbidden:
            description: Autenticado pero sin permisos suficientes
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        NotFound:
            description: El recurso solicitado no existe
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
...
        TooManyRequests:
            description: >
                Límite de peticiones excedido (global: 100/min por IP; endpoints
                de auth: 10/min; forgot-password: 3/min; contacto: 5/min;
                postulaciones: 20/min).
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
    schemas:
        SuccessEnvelope:
            type: object
            description: Sobre estándar de respuestas exitosas.
            properties:
                success:
                    type: boolean
                    example: true
                statusCode:
                    type: integer
                    example: 200
                message:
                    type: string
                    example: 'Operación exitosa'
                data:
                    description: Carga útil (objeto, array o null).
                meta:
                    $ref: '#/components/schemas/PaginationMeta'
            required: [success, statusCode, message, data]
        PaginationMeta:
            type: object
            description: Metadata de paginación (presente solo en listas paginadas).
            properties:
                totalItems: { type: integer, example: 145 }
                itemsPerPage: { type: integer, example: 10 }
                currentPage: { type: integer, example: 2 }
                totalPages: { type: integer, example: 15 }
                hasNextPage: { type: boolean, example: true }
                hasPrevPage: { type: boolean, example: true }
                nextPageUrl:
                    type: string
                    nullable: true
                    example: '/api/products?limit=10&page=3'
                prevPageUrl:
                    type: string
                    nullable: true
                    example: '/api/products?limit=10&page=1'
        ErrorResponse:
            type: object
            description: Sobre estándar de errores (lo formatean los Exception Filters).
            properties:
                success: { type: boolean, example: false }
                statusCode: { type: integer, example: 400 }
                message:
                    type: string
                    example: 'La contraseña debe tener al menos 8 caracteres...'
                errors:
                    type: array
                    nullable: true
                    items: { type: string }
                timestamp: { type: string, format: date-time }
                path: { type: string, example: '/api/auth/activate' }
...
        PublicUser:
            type: object
            description: Proyección segura del usuario (sin datos sensibles).
            properties:
                id:
                    type: string
                    format: uuid
                    example: '550e8400-e29b-41d4-a716-446655440000'
                email:
                    type: string
                    example: 'juan.perez@example.com'
                name:
                    type: string
                    example: 'Juan'
                surname:
                    type: string
                    example: 'Perez'
        UserEntity:
            type: object
            description: >
                Entidad de usuario serializada (password excluido por @Exclude).
                En una cuenta PENDIENTE de activación, name y surname son null.
            allOf:
                - $ref: '#/components/schemas/PublicUser'
                - type: object
                  properties:
                      status:
                          type: string
                          enum: [pendiente, activo, desactivado]
                          example: activo
                          description: >
                              Estado derivado, calculado por el backend a partir
                              de isActive y de si la cuenta ya definió contraseña.
                              Usalo tal cual: no lo deduzcas de isActive, que vale
                              false tanto para una invitación sin aceptar como
                              para una cuenta dada de baja.
                      name:
                          type: string
                          nullable: true
                          example: 'Juan'
                      surname:
                          type: string
                          nullable: true
                          example: 'Perez'
                      isActive:
                          type: boolean
                          example: true
                      isEmailVerified:
                          type: boolean
                          example: true
                      roles:
                          type: array
                          items:
                              type: string
                              enum: [user, admin, rrhh]
                          example: ['user']
                      createdAt:
                          type: string
                          format: date-time
                      updatedAt:
                          type: string
                          format: date-time
...